move some constants around

+++ shaders/shader.hpp
@@ -69,7 +69,7 @@ private:
     bool _enable_lightmap : 1 = false;
     Int _sampler = 0, _real_sampler;
 
-    enum : Int {
+    enum : uint8_t {
         ScaleUniform = 0, OffsetUniform = 1, TintUniform = 2,
         EnableLightmapUniform = 3,
         SamplerUniform = 4, LightmapSamplerUniform = 5,
